+++ backend/pkg/api/routes.go
@@ -70,6 +70,8 @@ func (api *API) routes() *chi.Mux {
 				r.Get("/topics-configs", api.handleGetTopicsConfigs())
 				r.Get("/topics-offsets", api.handleGetTopicsOffsets())
 				r.Get("/topics", api.handleGetTopics())
+				r.Delete("/topics/{topicName}", api.handleDeleteTopic())
+				r.Delete("/topics/{topicName}/records", api.handleDeleteTopicRecords())
 				r.Get("/topics/{topicName}/partitions", api.handleGetPartitions())
 				r.Get("/topics/{topicName}/configuration", api.handleGetTopicConfig())
 				r.Get("/topics/{topicName}/consumers", api.handleGetTopicConsumers())