+++ frontend/src/state/restInterfaces.ts
@@ -4,7 +4,7 @@ export interface ApiError {
 }
 
 
-export const TopicActions = ['seeTopic', 'viewPartitions', 'viewMessages', 'useSearchFilter', 'viewConsumers', 'viewConfig'] as const;
+export const TopicActions = ['seeTopic', 'viewPartitions', 'viewMessages', 'useSearchFilter', 'viewConsumers', 'viewConfig', 'deleteTopic'] as const;
 export type TopicAction = 'all' | typeof TopicActions[number];
 
 export interface Topic {
@@ -71,6 +71,17 @@ export interface GetAllPartitionsResponse {
         error: string | null;
         partitions: Partition[];
     }[];
+}
+
+export interface DeleteRecordsResponseData {
+    topicName: string;
+    partitions: [
+        {
+            partitionId: number;
+            lowWaterMark: number;
+            error: string;
+        }
+    ];
 }
 
 
@@ -622,12 +633,19 @@ export interface SchemaDetailsResponse {
     schemaDetails: SchemaDetails
 }
 
+export enum SchemaType {
+    AVRO="AVRO",
+    JSON="JSON",
+    PROTOBUF= "PROTOBUF",
+}
+
 export interface SchemaDetails {
     string: string;
     schemaId: number;
     version: number;
     compatibility: string;
-    schema: Schema;
+    type: SchemaType;
+    schema: Schema | JsonSchema;
     registeredVersions: number[];
 }
 
@@ -637,6 +655,30 @@ export interface Schema {
     namespace: string;
     type: string;
     fields: SchemaField[];
+}
+
+export enum JsonFieldType {
+    STRING="string",
+    NUMBER="number",
+    INTEGER= "integer",
+    OBJECT="object",
+    ARRAY="array",
+    BOOLEAN="boolean",
+    NULL="null"
+}
+
+export interface JsonSchema {
+  $id: string;
+  type: JsonFieldType;
+  title: string;
+  description: string;
+  properties?: Record<string, JsonField>;
+}
+
+export interface JsonField {
+  type: string;
+  items?: JsonField;
+  properties?: Record<string, JsonField>;
 }
 
 export interface SchemaField {